Proxy origin check from core
